# internal/model/event.go
// Package model holds the canonical Go types for the v0.1 event schema
// (prototype/manual-dataset-schema.md). These types are the in-memory contract
// shared by the parser, normalizer, classifier, store, and read API.
//
// Nullability follows the schema: scalar fields that the schema marks
// "string|null" / "date|null" use *string so the absence of a value is
// distinguishable from the empty string. Required scalars use string. Nested
// objects (Venue, Scale) use pointers so they can be null. Arrays default to a
// non-nil empty slice when "may be empty" and are required (>=1) where the
// schema demands it.
package model

// SchemaVersion is the schema revision these types implement.
const SchemaVersion = "0.1"

// Venue is the nested venue object (schema: Venue object).
type Venue struct {
	VenueID *string `json:"venue_id"` // slug, e.g. "coex"; nullable
	Name    string  `json:"name"`     // display name
	Hall    *string `json:"hall"`     // hall/booth area if known; nullable
	City    string  `json:"city"`     // city
}

// Scale is the optional disclosed-scale object.
type Scale struct {
	Visitors   *int `json:"visitors"`
	Exhibitors *int `json:"exhibitors"`
}

// Actions are the founder/operator boolean action fields. Each defaults to
// false when unknown (and the dotted key is then listed in MissingFields).
type Actions struct {
	CanRegister       bool `json:"can_register"`
	CanExhibit        bool `json:"can_exhibit"`
	CanSponsor        bool `json:"can_sponsor"`
	HasMatchmaking    bool `json:"has_matchmaking"`
	HasStartupProgram bool `json:"has_startup_program"`
}

// Source is one provenance entry. The schema requires >=1 per event.
type Source struct {
	URL         string `json:"url"`
	Type        string `json:"type"`         // venue|organizer|association|press|social|aggregator
	Publisher   string `json:"publisher"`    //
	RetrievedAt string `json:"retrieved_at"` // ISO8601; EXCLUDED from ContentHash
}

// Event is the canonical v0.1 record, mirroring
// prototype/manual-dataset-schema.md field-for-field.
type Event struct {
	// --- Identity ---
	SchemaVersion string  `json:"schema_version"` // "0.1"
	EventID       string  `json:"event_id"`       // durable slug, stable forever
	SeriesID      *string `json:"series_id"`
	Name          string  `json:"name"`
	NameKo        *string `json:"name_ko"`
	NameEn        *string `json:"name_en"`
	Edition       *string `json:"edition"`

	// --- When / Where ---
	StartDate      *string `json:"start_date"`      // ISO date or null
	EndDate        *string `json:"end_date"`        // ISO date or null
	Timezone       *string `json:"timezone"`        // IANA tz
	DateConfidence string  `json:"date_confidence"` // high|medium|low
	Status         string  `json:"status"`          // scheduled|tentative|postponed|cancelled|ended
	Format         string  `json:"format"`          // onsite|hybrid|online
	Venue          *Venue  `json:"venue"`
	Country        string  `json:"country"` // ISO 3166-1 alpha-2

	// --- Classification ---
	Categories []string `json:"categories"` // >=1, taxonomy slugs
	Audience   []string `json:"audience"`   // may be empty
	Scale      *Scale   `json:"scale"`

	// --- Founder / operator action fields ---
	Actions              Actions `json:"actions"`
	RegisterURL          *string `json:"register_url"`
	ExhibitURL           *string `json:"exhibit_url"`
	RegistrationDeadline *string `json:"registration_deadline"`
	ExhibitorDeadline    *string `json:"exhibitor_deadline"`
	CostHint             string  `json:"cost_hint"` // free|paid|mixed|unknown

	// Summary is a SHORT factual line generated by normalize from structured
	// fields only ("{name} — {start}~{end} @ {venue}, 주최 {organizer}", <=240
	// chars). It is NEVER copied from a source body (copyright). null when not
	// producible (and then "summary" is listed in MissingFields). Produced by
	// internal/normalize; not source-derived, so it is EXCLUDED from ContentHash
	// (it is not part of the frozen CanonicalFields projection).
	Summary *string `json:"summary"`

	// --- Provenance & freshness ---
	Sources        []Source `json:"sources"`         // >=1
	HomepageURL    *string  `json:"homepage_url"`    //
	LastCheckedAt  string   `json:"last_checked_at"` // EXCLUDED from ContentHash
	UpdateState    string   `json:"update_state"`    // new|unchanged|updated|conflicting (derived; EXCLUDED)
	Confidence     string   `json:"confidence"`      // high|medium|low
	MissingFields  []string `json:"missing_fields"`  // dotted paths
	AmbiguityNotes *string  `json:"ambiguity_notes"`

	// --- Curation meta (EXCLUDED from ContentHash) ---
	CuratedBy string `json:"curated_by"`
	CreatedAt string `json:"created_at"` // EXCLUDED from ContentHash
	UpdatedAt string `json:"updated_at"` // EXCLUDED from ContentHash

	// --- Operational ---
	ContentHash string `json:"content_hash,omitempty"` // computed; EXCLUDED from ContentHash input
	Excluded    bool   `json:"excluded"`               // non-industry; preserved, not source-semantic
	BatchID     string `json:"-"`                      // ingest batch; EXCLUDED from ContentHash
}

// ChangeLog is one field-level change record (the change feed row).
type ChangeLog struct {
	ID        int64   `json:"id"`
	EventID   string  `json:"event_id"`
	FieldPath string  `json:"field_path"` // dotted path, e.g. "status"
	OldValue  *string `json:"old_value"`
	NewValue  *string `json:"new_value"`
	ChangedAt string  `json:"changed_at"` // ISO8601
	BatchID   *string `json:"batch_id"`
}

// RawSnapshot is the last raw fetched payload for an event (provenance / reparse).
type RawSnapshot struct {
	EventID     string  `json:"event_id"`
	SourceID    *string `json:"source_id"`
	URL         *string `json:"url"`
	ContentType *string `json:"content_type"`
	Body        []byte  `json:"body"`
	ContentHash *string `json:"content_hash"`
	RetrievedAt string  `json:"retrieved_at"` // ISO8601
}